feat: Allow passing a custom ErrorMapper per request
This introduces the ability to provide a specific `ErrorMapper` for each individual network request.

If no `ErrorMapper` is provided for a request, the client will fall back to using the globally configured `ErrorMapper`. This allows for more flexible and granular error handling by enabling custom error parsing on a per-API-call basis.